`PwParser`: Keep scheduler parser error unless more specific available
When a `CalcJob` is retrieved, the engine will first call the parser of
the scheduler plugin if one is implemented. If that detects a problem
and returns an exit code, it will be set on the `CalcJobNode`. The
`PwParser` is updated to check for the presence of this exit code and
only override it if a more specific exit code can be provided.

Note that we have to manually recreate the `ExitCode` and return it
because returning `None` is currently equivalent to returning
`ExitCode(0)`.