`PwCalculation`: Add `ERROR_IONIC_INTERRUPTED_PARTIAL_TRAJECTORY` code
There are a number of reasons that a job can fail that will interrupt a
calculation, some of which may due to transient reasons. Typical
examples are parsed by the scheduler such a node failure or an out of
walltime error.

Restarting in this cases may be beneficial in the case of a relax
calculation where the partially computed and parsed trajectory can be
used to restart from. Here this case is parsed and when detected the
exit code `ERROR_IONIC_INTERRUPTED_PARTIAL_TRAJECTORY` is returned.